refactor deployment validator
